+++ src/main/java/org/opensearch/timeseries/rest/handler/IndexJobActionHandler.java
@@ -99,7 +99,7 @@ public abstract class IndexJobActionHandler<IndexType extends Enum<IndexType> & 
     private final ActionType<StopConfigResponse> stopConfigAction;
     protected final NodeStateManager nodeStateManager;
 
-    /**
+    /** 
      * Constructor function.
      *
      * @param client                  ES node client that executes actions on the local node